initial split in modules
