+++ ElasticSwift.xcodeproj/xcshareddata/xcschemes/ElasticSwift-Package.xcscheme
@@ -14,9 +14,65 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
+               BlueprintIdentifier = "ElasticSwift::ElasticSwiftCore"
+               BuildableName = "ElasticSwiftCore.framework"
+               BlueprintName = "ElasticSwiftCore"
+               ReferencedContainer = "container:ElasticSwift.xcodeproj">
+            </BuildableReference>
+         </BuildActionEntry>
+         <BuildActionEntry
+            buildForTesting = "YES"
+            buildForRunning = "YES"
+            buildForProfiling = "YES"
+            buildForArchiving = "YES"
+            buildForAnalyzing = "YES">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "ElasticSwift::ElasticSwiftNetworking"
+               BuildableName = "ElasticSwiftNetworking.framework"
+               BlueprintName = "ElasticSwiftNetworking"
+               ReferencedContainer = "container:ElasticSwift.xcodeproj">
+            </BuildableReference>
+         </BuildActionEntry>
+         <BuildActionEntry
+            buildForTesting = "YES"
+            buildForRunning = "YES"
+            buildForProfiling = "YES"
+            buildForArchiving = "YES"
+            buildForAnalyzing = "YES">
+            <BuildableReference
+               BuildableIdentifier = "primary"
                BlueprintIdentifier = "ElasticSwift::ElasticSwift"
                BuildableName = "ElasticSwift.framework"
                BlueprintName = "ElasticSwift"
+               ReferencedContainer = "container:ElasticSwift.xcodeproj">
+            </BuildableReference>
+         </BuildActionEntry>
+         <BuildActionEntry
+            buildForTesting = "YES"
+            buildForRunning = "YES"
+            buildForProfiling = "YES"
+            buildForArchiving = "YES"
+            buildForAnalyzing = "YES">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "ElasticSwift::ElasticSwiftDSL"
+               BuildableName = "ElasticSwiftDSL.framework"
+               BlueprintName = "ElasticSwiftDSL"
+               ReferencedContainer = "container:ElasticSwift.xcodeproj">
+            </BuildableReference>
+         </BuildActionEntry>
+         <BuildActionEntry
+            buildForTesting = "YES"
+            buildForRunning = "YES"
+            buildForProfiling = "YES"
+            buildForArchiving = "YES"
+            buildForAnalyzing = "YES">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "ElasticSwift::ElasticSwiftCodableUtils"
+               BuildableName = "ElasticSwiftCodableUtils.framework"
+               BlueprintName = "ElasticSwiftCodableUtils"
                ReferencedContainer = "container:ElasticSwift.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
@@ -26,18 +82,7 @@
       buildConfiguration = "Debug"
       selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
-      shouldUseLaunchSchemeArgsEnv = "YES"
-      codeCoverageEnabled = "YES"
-      onlyGenerateCoverageForSpecifiedTargets = "YES">
-      <CodeCoverageTargets>
-         <BuildableReference
-            BuildableIdentifier = "primary"
-            BlueprintIdentifier = "ElasticSwift::ElasticSwift"
-            BuildableName = "ElasticSwift.framework"
-            BlueprintName = "ElasticSwift"
-            ReferencedContainer = "container:ElasticSwift.xcodeproj">
-         </BuildableReference>
-      </CodeCoverageTargets>
+      shouldUseLaunchSchemeArgsEnv = "YES">
       <Testables>
          <TestableReference
             skipped = "NO">
