+++ Sources/ElasticSwift/Requests/DeleteRequest.swift
@@ -9,6 +9,7 @@
 import Foundation
 import Logging
 import NIOHTTP1
+import ElasticSwiftCore
 
 //MARK:- Delete Request Builder
