+++ Sources/ElasticSwift/Requests/SearchRequest.swift
@@ -9,6 +9,8 @@
 import Foundation
 import Logging
 import NIOHTTP1
+import ElasticSwiftCore
+import ElasticSwiftDSL
 
 //MARK:- Search Request Builder
