+++ Sources/ElasticSwift/Serialization/ResponseConverters.swift
@@ -6,6 +6,8 @@
 //
 
 import Foundation
+import ElasticSwiftCore
+import ElasticSwiftNetworking
 
 public typealias HTTPResponseHandler = (_ result: Result<HTTPResponse, Error>) -> Void
 public typealias ResultCallback<T: Codable> = (_ result: Result<T, Error>) -> Void
