+++ Sources/ElasticSwiftCore/Query.swift
