+++ Sources/ElasticSwiftCore/QueryBuilder.swift
