+++ Sources/ElasticSwiftDSL/Queries/Builders/FullTextBuilders.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import ElasticSwiftCore
 
 // MARK:- Builder for Match query.
