+++ Sources/ElasticSwiftDSL/Queries/Builders/GeoQueryBuilders.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import ElasticSwiftCore
 
 
 // MARK:- GeoShape Query Builder
