+++ Sources/ElasticSwiftDSL/Queries/Builders/JoiningQueryBuilders.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import ElasticSwiftCore
 
 // MARK:- Nested Query Builder
