+++ Sources/ElasticSwiftDSL/Queries/Builders/ScoreFunctionBuilders.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import ElasticSwiftCore
 
 public class ScoreFunctionBuilders {
