+++ Sources/ElasticSwiftDSL/Queries/Builders/SpanQueryBuilders.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import ElasticSwiftCore
 
 // MARK:- Span Term Query Builder
