+++ Sources/ElasticSwiftDSL/Queries/Builders/SpecializedQueryBuilders.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import ElasticSwiftCore
 
 // MARK:- MoreLikeThis Query Builder
