+++ Sources/ElasticSwiftDSL/Queries/FullTextQueries.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import ElasticSwiftCore
 
 // MARK:- Match Query.
