+++ Sources/ElasticSwiftDSL/Queries/GeoQueries.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import ElasticSwiftCore
 
 
 // MARK:- GeoShape Query
