+++ Sources/ElasticSwiftDSL/Queries/MatchAllQuery.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import ElasticSwiftCore
 
 // MARK:-  Match all Query
