+++ Sources/ElasticSwiftDSL/Queries/SpanQueries.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import ElasticSwiftCore
 
 // MARK:- Span Term Query
